Changed size_t to int

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -7,7 +7,7 @@
 size_t print_listint(const listint_t *h)
 {
 	const listint_t *p;
-	size_t counter = 0;
+	int counter = 0;
 
 	p = h;
 	while (p != NULL)
